requirements.txt
twitter~=1.18.0
python-twitter >= 3.4.1
requests~=2.9.1
kafka-python~=2.0.1
